Add td::TypeRegistry
This is useful for having a "perfect hash function" for types known at
link-time but not at compile-time. It works by defining a templated
function that, on the first entry, adds 1 to a global counter and saves
the value (via a static variable) and then returns the saved value on
subsequent entries.