Implement null consensus
This provides an alternative implementation of IValidatorGroup and can
be used when all nodes are trusted. The consensus itself is a debug tool
resistant to packet losses (TODO) but not to node protocol violations.

Actors that spawn with ConsensusBus will be reused for Alpenglow
consensus implementation in the future commits.